unify serializer for diff and saving snapshots
