+++ JestDotnet/JestDotnet/JestDotnet.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <TargetFrameworks>net6.0;net47;net50;netstandard2.0;netstandard2.1</TargetFrameworks>
-        <Version>1.4.0</Version>
+        <Version>1.4.1</Version>
         <Authors>Tomas Bruckner</Authors>
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
         <RepositoryUrl>https://github.com/tomasbruckner/jest-dotnet</RepositoryUrl>
